requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
flask==3.1.0
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
gunicorn 
python-dotenv==1.0.1
google-generativeai==0.8.3